requirements.txt
langchain
langchain-groq
python-dotenv
langchain_core
langchain_community
streamlit
langchain-chroma
requests